pyproject.toml
[build-system]
requires = ["hatchling>=1.27,<2"]
build-backend = "hatchling.build"

[project]
name = "zip2telegraph-bot"
version = "0.1.0"
description = "Telegram bot that converts ZIP image bundles into Telegraph pages."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiogram>=3.26,<4",
  "aiosqlite>=0.22,<0.23",
  "httpx>=0.28,<0.29",
  "Pillow>=11,<12",
]

[project.optional-dependencies]
dev = [
  "pytest>=9,<10",
  "pytest-asyncio>=1,<2",
]

[project.scripts]
zip2telegraph-bot = "zip2telegraph_bot.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/zip2telegraph_bot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
